+++ packages/svg/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @glissade/svg
 
+## 0.51.0-pre.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @glissade/core@0.51.0-pre.0
+  - @glissade/scene@0.51.0-pre.0
+
 ## 0.50.0
 
 ### Patch Changes