+++ packages/umbrella/CHANGELOG.md
@@ -1,5 +1,14 @@
 # glissade
 
+## 0.51.0-pre.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @glissade/core@0.51.0-pre.0
+  - @glissade/player@0.51.0-pre.0
+  - @glissade/scene@0.51.0-pre.0
+
 ## 0.50.0
 
 ### Patch Changes